Summary
ok so basically, the only way i could figure out how to put new buttons was to make a custom toolbar. This means some of the styling you added got all messed up, I was able to fix most of it, but some of the styling is still wonky. I’m gonna take a break rn, but I’ll try and fix it later.